Get movement but it's not correct
