+++ world.c
@@ -16,11 +16,12 @@
 #include <GL/glu.h>
 
 #include "draw_robot.h"
+#include "cube_floor.h"
 
 #define MAX_BUILDINGS 6
 #define BLOCKS_ROWS 20
 #define BLOCKS_COLUMNS 20
-#define BLOCKS_SZ 5.0
+#define BLOCKS_SZ 6.0
 #define GAPS_SZ 1.0
 
 static float eyeX =5.0f, eyeY = 5.0f, eyeZ = -1.0f; // for visualization, delete later.
@@ -47,234 +48,10 @@ struct block
 
 struct block city[BLOCKS_ROWS * BLOCKS_COLUMNS];
 
-// New
+extern int X_POS; 
+extern int Y_POS;
+extern int Z_POS;
 
-int BLOCK_LENGTH = 2;
-
-// This is the default position 
-// of the robot. 
-float X_POS = 3.5; 
-float Y_POS = 0.0;
-float Z_POS = 0.0;
-
-enum facing
-{
-    FACE_FORWARD,
-    FACE_RIGHT,
-    FACE_BACK,
-    FACE_LEFT
-};
-
-int FACING_STATE = FACE_FORWARD;
-
-enum alert_camera_views
-{
-	// Each of these descriptions are 
-	// relative to the  robots center.
-	FRONT_LEFT, // "F8"
-	FRONT_RIGHT, // "F7"
-	BACK_LEFT, // "F5"
-	BACK_RIGHT, // "F6"
-	FRONT_LEFT_FAR, // "F9"
-	FRONT_RIGHT_FAR, // "F10"
-	BACK_LEFT_FAR, // "F11"
-	BACK_RIGHT_FAR, // "F12"
-	DEFAULT // "No function key pressed"
-};
-int ALT_CAMERA_STATE = DEFAULT;
-
-float *get_camera_offset()
-{
-    // Return the offset (array of x, y, z) from the
-    // robot center to the camera 'eye' coordinates. 
-    static float xyz[3];
-
-    xyz[1] = Y_POS + 2;
-
-    switch (FACING_STATE)
-    {
-    case FACE_FORWARD:
-        xyz[0] = X_POS;
-        xyz[2] = Z_POS - 5;
-
-        switch (ALT_CAMERA_STATE)
-        {
-        case BACK_LEFT: 
-            xyz[0]+=5;
-			break;
-        case BACK_LEFT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;
-		case BACK_RIGHT: 
-            xyz[0]-=5;
-			break;
-        case BACK_RIGHT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;            
-        case FRONT_LEFT: 
-            xyz[0]+=5;
-            xyz[2]+=10;
-			break;
-        case FRONT_LEFT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;              
-		case FRONT_RIGHT: 
-            xyz[0]-=5;
-            xyz[2]+=10;
-			break;
-        case FRONT_RIGHT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;             
-		default: 
-			break;
-        }
-        break;
-    case FACE_BACK:
-        xyz[0] = X_POS;
-        xyz[2] = Z_POS + 5;
-
-        switch (ALT_CAMERA_STATE)
-        {
-        case BACK_LEFT: 
-            xyz[0]-=5;
-			break;
-        case BACK_LEFT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;
-		case BACK_RIGHT: 
-            xyz[0]+=5;
-			break;
-        case BACK_RIGHT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;            
-        case FRONT_LEFT: 
-            xyz[0]-=5;
-            xyz[2]-=10;
-			break;
-        case FRONT_LEFT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;              
-		case FRONT_RIGHT: 
-            xyz[0]+=5;
-            xyz[2]-=10;
-			break;
-        case FRONT_RIGHT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;             
-		default: 
-			break;
-        }
-        break;
-    case FACE_RIGHT: 
-        xyz[0] = X_POS + 5; 
-        xyz[2] = Z_POS;
-
-        switch (ALT_CAMERA_STATE)
-        {
-        case BACK_LEFT: 
-            xyz[2]+=5;
-			break;
-        case BACK_LEFT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;
-		case BACK_RIGHT: 
-            xyz[2]-=5;
-			break;
-        case BACK_RIGHT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;            
-        case FRONT_LEFT: 
-            xyz[0]-=10;
-            xyz[2]+=5;
-			break;
-        case FRONT_LEFT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;              
-		case FRONT_RIGHT: 
-            xyz[0]-=10;
-            xyz[2]-=5;
-			break;
-        case FRONT_RIGHT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;             
-		default: 
-			break;
-        }
-        break;        
-    case FACE_LEFT: 
-        xyz[0] = X_POS - 5;
-        xyz[2] = Z_POS;
-        switch (ALT_CAMERA_STATE)
-        {
-        case BACK_LEFT: 
-            xyz[2]-=5;
-			break;
-        case BACK_LEFT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;
-		case BACK_RIGHT: 
-            xyz[2]+=5;
-			break;
-        case BACK_RIGHT_FAR: 
-            xyz[0]-=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;            
-        case FRONT_LEFT: 
-            xyz[0]+=10;
-            xyz[2]-=5;
-			break;
-        case FRONT_LEFT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]-=100;
-			break;              
-		case FRONT_RIGHT: 
-            xyz[0]+=10;
-            xyz[2]+=5;
-			break;
-        case FRONT_RIGHT_FAR: 
-            xyz[0]+=100;
-            xyz[1]+=100;
-            xyz[2]+=100;
-			break;             
-		default: 
-			break;
-        }
-        break;         
-    default:
-        break;
-    }
-    return xyz;
-}
-
-// \New
 
 //returns random double between 0 & 1
 // used for city generation
@@ -462,35 +239,17 @@ void display(void)
 
 	//printf("%f %f %f \n",eyeX, eyeY, eyeZ);
    /* Viewing transformation */
-
-   float *p = get_camera_offset();
-   gluLookAt(p[0], p[1], p[2], X_POS, Y_POS, Z_POS, 0.0, 1.0, 0.0);
-
+  float *p = get_camera_offset();
+  gluLookAt(p[0], p[1], p[2], X_POS, Y_POS, Z_POS, 0.0, 1.0, 0.0);
   // The Main function that renders blocks & buildings
-  renderCity();
 
   glPushMatrix();
-  glTranslatef((GLfloat)X_POS, (GLfloat)Y_POS + 0.7, (GLfloat)Z_POS);
+  glTranslatef(3.0, 0.0, 0.0);
+  renderCity();
+  glPopMatrix();
 
-  switch (FACING_STATE)
-  // Rotate the robot according to the direction it is facing. 
-  {
-  case FACE_FORWARD:
-      glRotatef(0.0, 0.0, 1.0, 0.0);
-      break;
-  case FACE_RIGHT:
-      glRotatef(270.0, 0.0, 1.0, 0.0);
-      break;
-  case FACE_BACK:
-      glRotatef(180.0, 0.0, 1.0, 0.0);
-      break;
-  case FACE_LEFT:
-      glRotatef(90.0, 0.0, 1.0, 0.0);
-      break;
-  default:
-      break;
-  }
-
+  glPushMatrix();
+  glTranslatef((GLfloat)X_POS, (GLfloat)Y_POS, (GLfloat)Z_POS);
   draw_robot();
   glPopMatrix();
 
@@ -616,7 +375,10 @@ int main(int argc, char** argv)
    glutIdleFunc(display);
    glutReshapeFunc(reshape);
    glutMouseFunc(&mouse);
-   glutKeyboardFunc(&myCBKey);
+  //  glutKeyboardFunc(&myCBKey);
+  glutKeyboardFunc(keyboard);
+  glutSpecialFunc(special_keyboard);
+  glutSpecialUpFunc(special_keyboard_up);
    glutMainLoop();
    return 0;   /* We'll never be here.*/
 }
